requirements.txt
# Google Cloud Libraries
google-cloud-storage==2.14.0
google-cloud-documentai==2.20.0

# Data Processing
pandas==2.1.4

# Environment Management
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
